refactor(runtime): rename agent_loop.rs to agent_loop/mod.rs and extract types module
First step of agent_loop split (#3710 phase 2). Move the 13,853-LOC
`agent_loop.rs` to `agent_loop/mod.rs` and carve out the public-facing
type cluster — `LoopPhase`, `PhaseCallback`, `LoopOptions`,
`AgentLoopResult`, `ExperimentContext` (+ its `Default` impl) — into
`agent_loop/types.rs`.

mod.rs re-exports the five symbols via `pub use self::types::{...}` so
every external consumer keeps resolving the same paths
(`librefang_runtime::agent_loop::AgentLoopResult`,
`librefang_kernel::agent_loop::AgentLoopResult`, etc.).

types.rs imports only what it actually needs (`TokenUsage`,
`DecisionTrace`, `Arc`); fields that reference `crate::interrupt::*`,
`crate::aux_client::*`, or `librefang_types::config::*` keep their
full-path form so the new file pulls in no extra `use`s that would
collide with mod.rs's import block on a future merge.

The extraction is purely visibility-preserving: every type stays `pub`,
no internal call site changes. mod.rs: 13,853 -> 13,623 LOC (-230).
cargo check + cargo clippy --lib both clean.

Refs #3710
